pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cv-duplicate-detection-mdpi"
version = "0.1.0"
description = "MDPI Computer Vision Assessment"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "numpy>=2.2.6",
    "opencv-python-headless>=4.12.0.88",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "click>=8.1.7",
    "tensorflow>=2.15.0",
    "imagehash>=4.3.2",
    "scikit-learn>=1.8.0",
    "ruff>=0.14.10",
]

[project.scripts]
mdpi_assessment = "mdpi_assessment.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
]

[tool.ruff]
target-version = "py311"           
line-length = 88                    

[tool.ruff.lint]
select = ["E", "F", "W", "I", "C"]
ignore = ["E501"]
